requirements.txt
aiogram==3.7.0
google-generativeai==0.7.2
